# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.QuizHistoryResponse:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      submissions:
        items:
          $ref: '#/definitions/dto.QuizSubmissionResponse'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.QuizListResponse:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      quizzes:
        items:
          $ref: '#/definitions/dto.QuizResponse'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.QuizQuestionResponse:
    properties:
      id:
        type: string
      options:
        items:
          type: string
        type: array
      order_num:
        type: integer
      question:
        type: string
    type: object
  dto.QuizResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      questions:
        description: 列表中不包含，詳情中包含
        items:
          $ref: '#/definitions/dto.QuizQuestionResponse'
        type: array
      title:
        type: string
    type: object
  dto.QuizSubmissionRequest:
    properties:
      answers:
        additionalProperties:
          type: integer
        description: question_id -> option_index
        type: object
      quiz_id:
        type: string
    required:
    - answers
    - quiz_id
    type: object
  dto.QuizSubmissionResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      id:
        type: string
      quiz_title:
        type: string
      result:
        type: string
      score:
        type: integer
    type: object
...
  vo.ErrorResponse:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      error:
        type: string
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  vo.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
host: mindhelp.onrender.com
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: MindHelp 心理健康支援應用程式後端 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MindHelp Backend API
  version: "1.0"
paths:
...
  /quizzes:
    get:
      consumes:
      - application/json
      description: 獲取可用的心理測驗列表
      parameters:
      - description: 測驗類別
        in: query
        name: category
        type: string
      - default: 1
        description: 頁碼
        in: query
        name: page
        type: integer
      - default: 10
        description: 每頁數量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.QuizListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: 獲取測驗列表
      tags:
      - quiz
  /quizzes/{id}:
    get:
      consumes:
      - application/json
      description: 獲取測驗的完整內容和題目
      parameters:
      - description: 測驗ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.QuizResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      summary: 獲取測驗詳情
      tags:
      - quiz
  /quizzes/{id}/submit:
    post:
      consumes:
      - application/json
      description: 提交測驗答案並獲取結果
      parameters:
      - description: 測驗ID
        in: path
        name: id
        required: true
        type: string
      - description: 測驗答案
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.QuizSubmissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.QuizSubmissionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 提交測驗答案
      tags:
      - quiz
...
  /users/me/quiz_history:
    get:
      consumes:
      - application/json
      description: 獲取使用者的測驗完成歷史
      parameters:
      - default: 1
        description: 頁碼
        in: query
        name: page
        type: integer
      - default: 10
        description: 每頁數量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.QuizHistoryResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 獲取測驗歷史
      tags:
      - quiz
...
securityDefinitions:
  BearerAuth:
    description: 請輸入 "Bearer " 加上 JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"